+++ packages/framework/src/data/index.ts
@@ -1,7 +1,6 @@
 import type * as fetcher from "./fetcher";
 
 export * from "./fetcher";
-export * from "./dateFormatter";
 
 export interface Response {
   body?: string | object;
