+++ packages/framework/src/hooks/index.ts
@@ -6,4 +6,4 @@ export * from "./useTemplateData";
 export * from "./useWidgetId";
 export * from "./useEnsembleStorage";
 export * from "./useEvaluate";
-export * from "../data/dateFormatter";
+export * from "../date/dateFormatter";
