+++ packages/framework/src/index.ts
@@ -3,5 +3,6 @@ export * from "./loader";
 export * from "./parser";
 export * from "./hooks";
 export * from "./data";
+export * from "./date";
 export * from "./state";
 export { evaluate } from "./evaluate";
